Allow Debian installation
Update installation scripts to support Debian.  Since Ubuntu is a Debian
derivative, most of the work required adding the "Debian" string to
switches that looked for Ubuntu.

The one significant change was that sudo was required to call "brctl
show".  By adding this call to the etc/create_bridge.sh script,
everything worked as expected and allowed the test.sh file to run
properly.