Conan quick start
I've added a suggested quick start guide for the service- and kernel workflows to the top of the readme now. I haven't updated the rest of the readme yet, but I'm wondering if some of @staiyeba's details should be moved to the external docs or if we want to keep them here. @MagnusS let me know what you think. @KristianJerpetjon great if you can verify that the quick-start guide is correct and also works for linux distros.